GCC 4.9 build fix; bugfix; fixed a strange crash on application exit on Linux by refactoring CIconProvider into pointer-to-static-impl

+++ file-commander-core/src/ccontroller.cpp
@@ -387,5 +387,5 @@ int CController::currentDiskIndex(Panel p) const
 			return i;
 	}
 
-	return std::numeric_limits<size_t>::max();
+	return std::numeric_limits<int>::max();
 }
